# api/swagger.yaml
swagger: "2.0"
info:
  description: "An unofficial Tallinn Transportation API. Fetches data from the GTFS of [peatus.ee](peatus.ee/gtfs)."
  version: "1.0.0"
  title: "Tallinn Transport API"
  contact:
    email: "write@andreasvirkus.me"
  license:
    name: "MIT"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "tallinn-transport-api.herokuapp.com"
basePath: "/api/v1"
...
paths:
...
  /route/{routeId}:
    get:
      tags:
      - "route"
      summary: "Find route by ID"
      description: "Returns a single route"
      operationId: "getRoute"
      produces:
      - "application/json"
      parameters:
      - name: "routeId"
        in: "path"
        description: "ID of route to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Route not found"
...